fix: move analytics middleware
